Refactor validators, data fetching, main runtime, and IBKR reconciliation into owned modules with updated tests/docs, sanitized analysis-index logging, and full-suite coverage preserved
