+++ tests/ibkr/test_analysis_index.py
@@ -1,0 +1,6 @@
+"""Collected analysis-index tests extracted from reconciler cases."""
+
+from tests.ibkr.reconciler_cases import (
+    TestLoadLatestAnalyses,
+    TestParseScoresFromFinalDecision,
+)
