+++ tests/ibkr/test_opportunity_finder.py
@@ -1,0 +1,6 @@
+"""Collected opportunity-finder tests extracted from reconciler cases."""
+
+from tests.ibkr.reconciler_cases import (
+    TestCurrencyAccuracy,
+    TestPhase2CashBlockedCandidates,
+)
