+++ tests/ibkr/test_portfolio_health.py
@@ -1,0 +1,9 @@
+"""Collected portfolio-health tests extracted from reconciler cases."""
+
+from tests.ibkr.reconciler_cases import (
+    TestComputePortfolioHealth,
+    TestCorrelatedSellDetection,
+    TestCorrelatedSellDetectionWindow,
+    TestSecondaryRatioCorrelatedSell,
+    TestSellTypeTagging,
+)
