+++ tests/ibkr/test_position_evaluator.py
@@ -1,0 +1,3 @@
+"""Collected held-position evaluation tests extracted from reconciler cases."""
+
+from tests.ibkr.reconciler_cases import TestConcentration, TestReconcile
