+++ tests/ibkr/test_reconciliation_rules.py
@@ -1,0 +1,10 @@
+"""Collected reconciliation-rules tests extracted from reconciler cases."""
+
+from tests.ibkr.reconciler_cases import (
+    TestCheckStaleness,
+    TestCheckStopBreach,
+    TestCheckTargetHit,
+    TestExchangeFromPosition,
+    TestResolveFx,
+    TestStalenessDisplay,
+)
