+++ tests/ibkr/test_watchlist_evaluator.py
@@ -1,0 +1,7 @@
+"""Collected watchlist-evaluator tests extracted from reconciler cases."""
+
+from tests.ibkr.reconciler_cases import (
+    TestHeldPositionBlocksCandidate,
+    TestWatchlistSuffixDedup,
+    TestWatchlistVerdictRouting,
+)
